Add Fleeing state and corresponding changes to ai commander logic
I need to modify other existing states so that they're aware of the
fleeing state and know which can change to it, and if the can, when
they should do it.

As of now, ai commander will prioritize running away and using
ranged attacks when low on health. It will also cast it's attack but
wait until target get's in range to use it.