make actors ask target unit for valid actors, so that...
... they don't end up all targeting the same single actor.

The way it works now the unit will choose the first actor with
less targets on its back.

I'm not sure this is the best place for this, because we are filtering
actors by preference and then asking the target's unit for a valid
actor so that they don't end up all targeting the same actor.

I think ideally, we should refactor the preference filter to find a
valid unit, not a valid actor.

So it should be the "nearest" unit, or attacking "unit", or most
damaged "unit" and then ask the unit for a valid actor. But it was
a bigger refactor so I left it for later.