feature/base_for_unit_ai (#46)
* Add Debug visualizations node to Unit

This helps us see where is the unit, who are its actors, and
once there is a target, what other unit it is aiming at

* first version of Unit States and State Machine

* implement unit state dead

* refactor search_destroy state

* WIP Move Unit state files and reorganize actor states folders

* refactor how actor gets its states

This is so that both actor and unit can use similar logic to get their states

* refactor unit to use factory and ref data to get their own states

* make actors target other actors by unit

This limits actors to only target enemy targets from the designated unit,
or ally targets from any ally unit

* make actors ask target unit for valid actors, so that...

... they don't end up all targeting the same single actor.

The way it works now the unit will choose the first actor with
less targets on its back.

I'm not sure this is the best place for this, because we are filtering
actors by preference and then asking the target's unit for a valid
actor so that they don't end up all targeting the same actor.

I think ideally, we should refactor the preference filter to find a
valid unit, not a valid actor.

So it should be the "nearest" unit, or attacking "unit", or most
damaged "unit" and then ask the unit for a valid actor. But it was
a bigger refactor so I left it for later.